refactor: post jpaRepository 이름 변경
